+++ tour-service/src/main/java/site/thanhtungle/tourservice/model/dto/response/tour/TourResponseDTO.java
@@ -11,6 +11,7 @@ import site.thanhtungle.tourservice.model.dto.response.tourinclude.SimpleTourInc
 import site.thanhtungle.tourservice.model.entity.TourImage;
 import site.thanhtungle.tourservice.model.entity.TourItinerary;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
@@ -24,8 +25,8 @@ public class TourResponseDTO {
     private String duration;
     private Float ratingAverage;
     private Integer ratingQuantity;
-    private Float price;
-    private Float priceDiscount;
+    private BigDecimal price;
+    private BigDecimal priceDiscount;
     private String tourType;
     private Integer groupSize;
     private String summary;
